requirements.txt
absl-py==2.1.0
accelerate==0.26.1
aiohttp==3.9.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.22.1
bleach==6.1.0
blis==0.7.11
boto3==1.34.29
botocore==1.34.29
bpemb==0.3.4
bs4==0.0.2
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorclass==2.2.2
confection==0.1.4
conllu==4.5.3
contourpy==1.2.0
cyac==1.9
cycler==0.12.1
cymem==2.0.8
Cython==3.0.8
datasets==2.16.1
dateparser==1.2.0
DateTime==5.4
de-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.7.0/de_core_news_lg-3.7.0-py3-none-any.whl#sha256=1451e06dedb5c3bb096127bf7c2117c253265fdfcad29ee488ff57bec5f78eaa
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
docopt==0.6.2
executing==2.0.1
fastapi==0.105.0
fastjsonschema==2.19.1
filelock==3.13.1
flair==0.13.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
fuzzywuzzy==0.18.0
gast==0.5.4
gdown==5.0.1
gensim==4.3.2
git-lfs==1.6
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
htmldate==1.4.3
huggingface-hub==0.20.3
idna==3.6
install==1.3.5
interchange==2021.0.4
ipython==8.12.3
Janome==0.5.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpickle==3.0.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
Levenshtein==0.23.0
libclang==16.0.6
logger==1.4
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.2
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==10.2.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
names-dataset==3.1.0
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
neo4j==5.15.0
neomodel==5.2.1
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
pansi==2020.7.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip-upgrader==1.4.15
pipreqs==0.5.0
platformdirs==4.2.1
pptree==3.1
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py2neo==2021.2.4
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==23.12.11
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-Levenshtein==0.23.0
pytorch_revgrad==0.2.0
pytz==2023.3.post1
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==26.0.2
rapidfuzz==3.6.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
segtok==1.5.11
semver==3.0.2
sentence-transformers==2.3.0
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spikex==0.5.2
sqlitedict==2.1.0
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.16.1
termcolor==2.4.0
terminaltables==3.1.10
tf-cpu==0.0.0
thinc==8.2.2
threadpoolctl==3.2.0
tinycss2==1.3.0
tokenizers==0.13.3
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformer-smaller-training-vocab==0.3.3
transformers==4.25.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
Werkzeug==3.0.1
Wikipedia-API==0.6.0
wrapt==1.14.1
xgboost==2.0.3
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
zope.interface==6.1
